Refactor the queries in different files
