+++ dashboard/src/server/db/queries/index.ts
@@ -1,0 +1,28 @@
+export { countMetrics, getMetricsByTag, meanMetric } from "./metrics-table";
+export {
+  getAvgDocsPerQuery,
+  getResponseTimeTrend,
+  getSystemHealthStats,
+  getTokenUsageStats,
+} from "./rag-quality";
+export { topKTopics } from "./topic-counts";
+export type {
+  ActivityByDay,
+  HourlyActivity,
+  MetricsByTag,
+  MetricsQueryParams,
+  ResponseTimeTrend,
+  SearchTerm,
+  SystemHealthStats,
+  TokenUsageStats,
+  TopicCount,
+} from "./types";
+export {
+  getActivityByDay,
+  getHourlyActivityPattern,
+  getTotalActivityEvents,
+  getUniqueUsers,
+  getUserRoleDistribution,
+  meanSessionLength,
+} from "./user-activity";
+export { topKSearchTerms } from "./word-counts";
